Fixed up issue of medication times that were not selected to be Skipped or Taken not being registered as a missing medication time in Weekly View of Medication Log page. Also standardized horizontal padding for multiple metric pages, renamed ContentView appropriately to DashboardView for more clarity, and moved public Date extension into its own file to be accessed by multiple pages more clearly
