+++ Prevonto/Date.swift
@@ -1,0 +1,15 @@
+import SwiftUI
+
+// Properties defined for Date
+extension Date {
+    static func from(year: Int, month: Int, day: Int, hour: Int = 0, minute: Int = 0, second: Int = 0) -> Date {
+        var components = DateComponents()
+        components.year = year
+        components.month = month
+        components.day = day
+        components.hour = hour
+        components.minute = minute
+        components.second = second
+        return Calendar.current.date(from: components)!
+    }
+}
